requirements/requirements-local.txt
PyYAML>=6.0.2
huggingface_hub>=0.31.1
openai>=1.76.0
requests>=2.32.3
transformers>=4.46.0
torch>=2.6.0
safetensors>=0.5.3
